typing and interfaces implementation
