+++ types/types/course.types.ts
@@ -1,0 +1,13 @@
+
+export interface Course {
+_id: string;
+  name: string;
+  endpoint: string;
+  content: {
+    title: string;
+    description: string;
+    url: string;
+    idVideo: number;
+  }[];
+  id: number;
+}
